src/tools/nlp/txt_tokens/requirements.txt
tokenizer==3.4.5